site/requirements.txt
Django>=4.2,<5
django_compressor>=3
django-mptt>=0.13
django-registration-redux>=2.10
django-reversion>=5
django-social-share
django-sortedm2m>=3.1.0
django-impersonate
dmoj-wpadmin @ git+https://github.com/DMOJ/dmoj-wpadmin.git
lxml
lxml_html_clean
Pygments
mistune<2
social-auth-core==4.3.0
social-auth-app-django==5.0.0
pytz
django-statici18n
pika
ua-parser
pyyaml
jinja2
django_jinja>=2.5.0
llist
requests
django-fernet-fields @ git+https://github.com/DMOJ/django-fernet-fields.git
pyotp
qrcode[pil]
jsonfield @ git+https://github.com/DMOJ/jsonfield.git
pymoss
packaging<22
celery
ansi2html @ git+https://github.com/DMOJ/ansi2html.git
sqlparse
lupa
netaddr
webauthn<1
bleach[css]
django-admin-sortable2
icalendar
pymysql==1.0.2
cryptography
redis
# This is a celery dependency whose latest major version is breaking everything.
importlib-metadata<5
# REST API
djangorestframework>=3.14
